+++ nengo/connection.py
@@ -1,5 +1,4 @@
 import logging
-import weakref
 
 import numpy as np
 
@@ -10,7 +9,7 @@ from nengo.exceptions import ValidationError
 from nengo.learning_rules import LearningRuleType, LearningRuleTypeParam
 from nengo.node import Node
 from nengo.params import (Default, Unconfigurable, ObsoleteParam,
-                          BoolParam, FunctionParam)
+                          BoolParam, FunctionInfo, FunctionParam)
 from nengo.solvers import LstsqL2, SolverParam
 from nengo.synapses import Lowpass, SynapseParam
 from nengo.utils.compat import is_array_like, is_iterable, iteritems
@@ -32,7 +31,6 @@ class ConnectionLearningRuleTypeParam(LearningRuleTypeParam):
     """Connection-specific validation for learning rules."""
 
     def __set__(self, conn, rule):
-        conn._learning_rule = None
         super(ConnectionLearningRuleTypeParam, self).__set__(conn, rule)
 
     def validate_rule(self, conn, rule):
@@ -119,14 +117,16 @@ class ConnectionFunctionParam(FunctionParam):
 
     def __set__(self, conn, function):
         if function is None:
-            function_info = self.Info(function=None, size=None)
+            function_info = FunctionInfo(function=None, size=None)
+        elif isinstance(function, FunctionInfo):
+            function_info = function
         elif is_array_like(function):
             array = np.array(function, copy=False, dtype=np.float64)
             self.validate_array(conn, array)
-            function_info = self.Info(function=array, size=array.shape[1])
+            function_info = FunctionInfo(function=array, size=array.shape[1])
         elif callable(function):
-            function_info = self.Info(function=function,
-                                      size=self.determine_size(conn, function))
+            function_info = FunctionInfo(
+                function=function, size=self.determine_size(conn, function))
             self.validate_callable(conn, function_info)
         else:
             raise ValidationError("Invalid connection function type %r "
@@ -382,6 +382,10 @@ class Connection(NengoObject):
         since="v2.1.0",
         url="https://github.com/nengo/nengo/issues/632#issuecomment-71663849")
 
+    _param_init_order = [
+        'pre', 'post', 'synapse', 'transform', 'eval_points', 'function_info',
+        'solver', 'learning_rule_type']
+
     def __init__(self, pre, post, synapse=Default, function=Default,
                  transform=Default, solver=Default, learning_rule_type=Default,
                  eval_points=Default, scale_eval_points=Default,
@@ -442,22 +446,24 @@ class Connection(NengoObject):
     @property
     def learning_rule(self):
         """(LearningRule or iterable) Connectable learning rule object(s)."""
-        if self.learning_rule_type is not None and self._learning_rule is None:
-            types = self.learning_rule_type
-            if isinstance(types, dict):
-                self._learning_rule = type(types)()  # dict of same type
-                for k, v in iteritems(types):
-                    self._learning_rule[k] = LearningRule(self, v)
-            elif is_iterable(types):
-                self._learning_rule = [LearningRule(self, v) for v in types]
-            elif isinstance(types, LearningRuleType):
-                self._learning_rule = LearningRule(self, types)
-            else:
-                raise ValidationError(
-                    "Invalid type %r" % type(types).__name__,
-                    attr='learning_rule_type', obj=self)
+        if self.learning_rule_type is None:
+            return None
 
-        return self._learning_rule
+        types = self.learning_rule_type
+        if isinstance(types, dict):
+            learning_rule = type(types)()  # dict of same type
+            for k, v in iteritems(types):
+                learning_rule[k] = LearningRule(self, v)
+        elif is_iterable(types):
+            learning_rule = [LearningRule(self, v) for v in types]
+        elif isinstance(types, LearningRuleType):
+            learning_rule = LearningRule(self, types)
+        else:
+            raise ValidationError(
+                "Invalid type %r" % type(types).__name__,
+                attr='learning_rule_type', obj=self)
+
+        return learning_rule
 
     @property
     def post_obj(self):
@@ -514,7 +520,7 @@ class LearningRule(object):
     """
 
     def __init__(self, connection, learning_rule_type):
-        self._connection = weakref.ref(connection)
+        self._connection = connection
         self.learning_rule_type = learning_rule_type
 
     def __repr__(self):
@@ -525,10 +531,19 @@ class LearningRule(object):
         return "<LearningRule modifying %s with type %s>" % (
             self.connection, self.learning_rule_type)
 
+    def __eq__(self, other):
+        return (
+            self._connection is other._connection and
+            self.learning_rule_type == other.learning_rule_type)
+
+    def __hash__(self):
+        # +1 to avoid collision with ensemble
+        return hash(self._connection) + hash(self.learning_rule_type) + 1
+
     @property
     def connection(self):
         """(Connection) The connection modified by the learning rule."""
-        return self._connection()
+        return self._connection
 
     @property
     def error_type(self):