+++ nengo/exceptions.py
@@ -9,6 +9,10 @@ class NengoException(Exception):
     """
 
 
+class NengoWarning(Warning):
+    """Base class for Nengo warnings."""
+
+
 class ValidationError(NengoException, ValueError):
     """A ValueError encountered during validation of a parameter."""
 
@@ -98,3 +102,18 @@ class CacheIOError(NengoException, IOError):
 
 class TimeoutError(NengoException):
     """A timeout occurred while waiting for a resource."""
+
+
+class NotAddedToNetworkWarning(NengoWarning):
+    """A NengoObject has not been added to a network."""
+
+    def __init__(self, obj):
+        self.obj = obj
+        super(NotAddedToNetworkWarning, self).__init__()
+
+    def __str__(self):
+        return (
+            "{obj} was not added to the network. When copying objects, "
+            "use the copy method on the object instead of Python's copy "
+            "module. When unpickling objects, they have to be added to "
+            "networks manually.".format(obj=self.obj))